Voice-to-text/requirements.txt
fastapi
uvicorn
python-multipart

openai-whisper
reportlab

langchain
langchain-community
langchain-core
langgraph
langchain-google-genai

chromadb
sentence-transformers

pypdf
Pillow

google-generativeai
python-dotenv
requests
